requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2

langchain #core framework
langchain-community # extra integration
faiss-cpu # fast vector database to score the embedded data
langchain-huggingface  # connect hugging face models to perform embedding
langchain-text-splitters # to split the data into smaller chunk
sentence-transformers # pre_trained models to connect chunkx into vector
langchain-core # to handle documents chains of data etc.,